moved database stuff from mainlayout to dbcontext
