+++ my-app/src/pages/_app.js
@@ -2,6 +2,7 @@ import * as React from "react";
 import { ChakraProvider, extendTheme } from "@chakra-ui/react";
 import NavBar from '../components/NavBar';
 import { Web3Provider } from "@/contexts/Web3Context";
+import {DataBaseProvider} from "@/contexts/DataBaseContext";
 
 
 
@@ -14,12 +15,14 @@ const theme = extendTheme({
 
 function MyApp({ Component, pageProps }) {
   return (
+    <DataBaseProvider>
     <Web3Provider>
     <ChakraProvider theme={theme}>
       <NavBar />
       <Component {...pageProps} />
     </ChakraProvider>
     </Web3Provider>
+    </DataBaseProvider>
   );
 }
